Lots of changes and additions.
- Added documentation to A2SRulesResponse
- Refactored A2SUpdater to include updateA2S method, allowing for method reference lamba when scheduling A2S updates.
- Removes AdminCamTracker
- Added AdminListReader utility class. Is a first iteration that was lazily done.
- Added documentation to ChatCommands and implemented broadcast commands.
- Removed ChatLogger
- Changed documentation for ClassHelper to conform to javadoc standards.
- Moved config.json
- Modified ConfigLoader to read configuration at runtime.
- Added some logic to abstract Connector class
- Added DBLog plugin with a base skeleton. Can't be finished until MySQLConnector is finished.
- Modified DeployableDamagedEvent to use double data type for damage and health remaining instead of integer.
- Added some more things to DevelopmentTestPlugin
- Modified Event documentation to conform to javadoc.
- Modified EventEmitter to use logger level trace instead of debug for plugin instances.
- Added documentation to GlobalThreadPool
- Modified LogParser to correctly parse chain IDs and use double values for classes that changed to have fields use double.
- Added an "optimization" to LogTailer such that lines aren't send to the log parser if they shouldn't be, with the intent of reducing the load on LogParser
- Added MySQLConnector initialization in Main
- Implemented the following functionalities into MySQLConnector: connecting to server, creating tables if they don't exist, getting match, inserting playercount. The current match logic needs to be adjusted for this.
- Added documentation to NamedThreadFactory
- Added package-info.java for plugins package.
- Modified PlayerDamagedEvent to use double data type for damage instead of integer.
- Modified PlayerDiedEvent to use double data type for damage instead of integer.
- Changed documentation on PlayerListUpdatedEvent to conform to javadoc.
- Modified PlayerWoundedEvent to use double data type for damage instead of integer.
- Changed documention on RconImpl to conform to javadoc.
- Modified RconUpdater to split updating player list, squad list, and layer info into different methods, which are called when the scheduler executes.
- Minor changes to README.md
- Changed documentation on SquadListUpdatedEvent to conform to javadoc.
- Added the following to SquadServer: fields for tracking names to steam64ids, collections for players and squads, and admins; initializing admin fields by reading admin lists (first iteration, needs improvement,); implementation for receiving NEW_GAME, PLAYER_DISCONNECTED, ROUND_WINNER, STEAMID_CONNECTED, PLAYER_BANNED, PLAYER_KICKED, PLAYERLIST_UPDATED, POSSESSED_ADMIN_CAM, SQUADLIST_UPDATED, and UNPOSSESSED_ADMIN_CAM events; public methods to get players by steamid; retrieval of server ID; getters for most fields.
- Modified TailerService to experiment with reading log file.

+++ src/main/java/event/logparser/DeployableDamagedEvent.java
@@ -14,11 +14,11 @@ import java.util.Date;
  */
 public class DeployableDamagedEvent extends Event {
     private final String deployable;
-    private final Integer damage;
+    private final Double damage;
     private final String weapon;
     private final String playerSuffix;
     private final String damageType;
-    private final Integer healthRemaining;
+    private final Double healthRemaining;
 
     /**
      * Constructs a {@link DeployableDamagedEvent}.
@@ -33,8 +33,8 @@ public class DeployableDamagedEvent extends Event {
      * @param damageType the damage type done to the deployable
      * @param healthRemaining the amount of health remaining on the deployable
      */
-    public DeployableDamagedEvent(Date date, EventType type, Integer chainID, String deployable, Integer damage, String weapon,
-                                  String playerSuffix, String damageType, Integer healthRemaining){
+    public DeployableDamagedEvent(Date date, EventType type, Integer chainID, String deployable, Double damage, String weapon,
+                                  String playerSuffix, String damageType, Double healthRemaining){
         super(date, type, chainID);
         this.deployable = deployable;
         this.damage = damage;
@@ -58,7 +58,7 @@ public class DeployableDamagedEvent extends Event {
      *
      * @return the amount of damage done to the deployable
      */
-    public Integer getDamage() {
+    public Double getDamage() {
         return damage;
     }
 
@@ -94,7 +94,7 @@ public class DeployableDamagedEvent extends Event {
      *
      * @return the health reamining on the deployable
      */
-    public Integer getHealthRemaining() {
+    public Double getHealthRemaining() {
         return healthRemaining;
     }
 }
